Refactoring & Building System & Player
Start use namespaces (not complited)
Added base of building system (not complited)
Added "state machine" for class "Player" (not complited)

+++ Shooter-2.0/Assets/Scripts/Resource/Resource.cs
@@ -1,4 +1,5 @@
 using System;
+using TSI.Entities;
 using UnityEngine;
 
 public abstract class Resource : HealthEntity
@@ -13,7 +14,7 @@ public abstract class Resource : HealthEntity
 
     public ResourceType Type => _type;
     public int Endurance => _endurance;
-    public float Amount => _amount;
+    public int Amount => _amount;
 
     public abstract void Extract();
 }
